Source code V1.1
This is still the first version of the software, and this update separates the previous source code from the PDF file in four parts for easy download and operation. At the same time, V1. 0 has been corrected for some errors and defects, including syntax errors on the C+ + backend, logical sequence errors, thread safety issues, parameter binding errors, null value handling errors, and React the front-end component import is incomplete, the state definition is truncated, the API request logic is wrong, the form verification rules are incomplete, and the label is wrong. I hope this correction can make this software experience better.